requirements.txt
streamlit
ultralytics
opencv-python
pillow
numpy
pytesseract
pdf2image
pandas
